requirements/default.txt
boto
confypy
docopt
jinja2
paramiko
pyyaml
salt-cloud
salt